+++ pkg/ui/react-app/src/globals.ts
@@ -2,3 +2,5 @@ import jquery from 'jquery';
 
 (window as any).jQuery = jquery;
 (window as any).moment = require('moment');
+
+(window as any).THANOS_QUERY_URL = '';